+++ program/lib/Roundcube/session/memcache.php
@@ -136,7 +136,7 @@ class rcube_session_memcache extends rcube_session
      *
      * @return bool True on success, False on failure
      */
-    public function write($key, $vars)
+    protected function save($key, $vars)
     {
         if ($this->ignore_write) {
             return true;
@@ -161,7 +161,7 @@ class rcube_session_memcache extends rcube_session
      *
      * @return bool True on success, False on failure
      */
-    public function update($key, $newvars, $oldvars)
+    protected function update($key, $newvars, $oldvars)
     {
         $ts = microtime(true);