+++ program/lib/Roundcube/session/memcached.php
@@ -135,7 +135,7 @@ class rcube_session_memcached extends rcube_session
      *
      * @return bool True on success, False on failure
      */
-    public function write($key, $vars)
+    protected function save($key, $vars)
     {
         if ($this->ignore_write) {
             return true;
@@ -160,7 +160,7 @@ class rcube_session_memcached extends rcube_session
      *
      * @return bool True on success, False on failure
      */
-    public function update($key, $newvars, $oldvars)
+    protected function update($key, $newvars, $oldvars)
     {
         $ts = microtime(true);